# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My swager Doc",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Development Server"
        }
    ],
    "paths": {
...
        "/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get info about own profile",
                "operationId": "267ab80ac2c6b3799a49d89a30cb6363",
                "responses": {
                    "200": {
                        "description": "prepare order",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Jhon"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "jhon@mail.ru"
                                                },
                                                "phone_number": {
                                                    "type": "string",
                                                    "example": "89617756027"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "locations": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "New York"
                                                    },
                                                    "street": {
                                                        "type": "string",
                                                        "example": "Lenina"
                                                    },
                                                    "house_number": {
                                                        "type": "string",
                                                        "example": "25b"
                                                    },
                                                    "floor": {
                                                        "type": "integer",
                                                        "example": 15
                                                    },
                                                    "apartment": {
                                                        "type": "integer",
                                                        "example": 151
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Ooops! Looks like something went wrong."
                                        },
                                        "message": {
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/location": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get info about own profile",
                "operationId": "49bb60d4f6b3aca210e68e59041fd9c9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "required": [
                                            "city",
                                            "street",
                                            "house_number"
                                        ],
                                        "properties": {
                                            "city": {
                                                "type": "string",
                                                "example": "New York"
                                            },
                                            "street": {
                                                "type": "string",
                                                "example": "Lenina"
                                            },
                                            "house_number": {
                                                "type": "string",
                                                "example": "25b"
                                            },
                                            "floor": {
                                                "type": "integer",
                                                "example": 15
                                            },
                                            "apartment": {
                                                "type": "integer",
                                                "example": 151
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "city": {
                                            "type": "string",
                                            "example": "New York"
                                        },
                                        "street": {
                                            "type": "string",
                                            "example": "Lenina"
                                        },
                                        "house_number": {
                                            "type": "string",
                                            "example": "25b"
                                        },
                                        "floor": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "apartment": {
                                            "type": "integer",
                                            "example": 151
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/location/{location}": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get info about own profile",
                "operationId": "b899806311a7b59c08e9eddaf30f24d3",
                "parameters": [
                    {
                        "name": "location",
                        "in": "path",
                        "description": "location id",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "city": {
                                                "type": "string",
                                                "example": "New York"
                                            },
                                            "street": {
                                                "type": "string",
                                                "example": "Lenina"
                                            },
                                            "house_number": {
                                                "type": "string",
                                                "example": "25b"
                                            },
                                            "floor": {
                                                "type": "integer",
                                                "example": 15
                                            },
                                            "apartment": {
                                                "type": "integer",
                                                "example": 151
                                            },
                                            "deleted": {
                                                "type": "boolean",
                                                "example": false
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "city": {
                                            "type": "string",
                                            "example": "New York"
                                        },
                                        "street": {
                                            "type": "string",
                                            "example": "Lenina"
                                        },
                                        "house_number": {
                                            "type": "string",
                                            "example": "25b"
                                        },
                                        "floor": {
                                            "type": "integer",
                                            "example": 15
                                        },
                                        "apartment": {
                                            "type": "integer",
                                            "example": 151
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Ooops! Looks like something went wrong."
                                        },
                                        "message": {
                                            "example": "Not Found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "if location was deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Ooops! Looks like something went wrong."
                                        },
                                        "message": {
                                            "example": "Forbiden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "example": "Ooops! Looks like something went wrong."
                                        },
                                        "message": {
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the following format: `Bearer {token}`",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}